+++ backend/src/main/java/nl/vng/diwi/dal/entities/PlanCategory.java
@@ -1,0 +1,24 @@
+package nl.vng.diwi.dal.entities;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import nl.vng.diwi.dal.GenericRepository;
+import nl.vng.diwi.dal.entities.superclasses.IdSuperclass;
+
+import java.util.List;
+
+@Entity
+@Table(name = "plan_category", schema = GenericRepository.VNG_SCHEMA_NAME)
+@Getter
+@Setter
+@NoArgsConstructor
+public class PlanCategory extends IdSuperclass {
+
+    @OneToMany(mappedBy="planCategory", fetch = FetchType.LAZY)
+    private List<PlanCategoryState> states;
+}